+++ x64barebones/Kernel/syscalls/sys_consume_stdin.c
@@ -2,6 +2,6 @@
 #include <keyboard.h>
 
 
-unsigned int sys_consume_stdin(char * buf, unsigned int count) {
+uint64_t sys_consume_stdin(char * buf, unsigned int count) {
 	return consume_kb_buffer(buf, count);
 }
